Component re-org
