+++ frontend/components/App/index.ts
@@ -1,0 +1,2 @@
+import App from './App'
+export default App
