+++ frontend/components/Checkbox/index.ts
@@ -1,0 +1,2 @@
+import Checkbox from './Checkbox'
+export default Checkbox
