+++ frontend/components/Footer/Footer.tsx
@@ -1,0 +1,27 @@
+import * as React from 'react'
+import Icon from '../Icon'
+import AlgoliaLogo from './AlgoliaLogo'
+import styles from './Footer.styles'
+
+const githubLink = 'https://github.com/jschr/vscodethemes'
+const twitterLink = `https://twitter.com/intent/tweet?text=${encodeURIComponent(
+  'Preview themes from the VSCode marketplace',
+)}&via=_jschr&url=${encodeURIComponent(
+  'https://vscodethemes.com',
+)}&hashtags=vscode`
+
+const Footer: React.SFC<{}> = () => (
+  <div className={styles.footer}>
+    <div className={styles.links}>
+      <a className={styles.link} href={githubLink}>
+        <Icon className={styles.icon} icon="github" />View on Github
+      </a>
+      <a className={styles.link} href={twitterLink} target="_blank">
+        <Icon className={styles.icon} icon="twitter" /> Share on Twitter
+      </a>
+    </div>
+    <AlgoliaLogo />
+  </div>
+)
+
+export default Footer
