+++ frontend/components/Footer/index.ts
@@ -1,0 +1,2 @@
+import Footer from './Footer'
+export default Footer
