+++ frontend/components/Heading/index.ts
@@ -1,0 +1,2 @@
+import Heading from './Heading'
+export default Heading
