+++ frontend/components/Icon/index.ts
@@ -1,0 +1,3 @@
+import Icon, { Icons } from './Icon'
+export default Icon
+export { Icons }
