+++ frontend/components/Input/index.ts
@@ -1,0 +1,2 @@
+import Input from './Input'
+export default Input
