+++ frontend/components/Pagination/index.ts
@@ -1,0 +1,5 @@
+import PageLink from './PageLink'
+import Pagination from './Pagination'
+
+export default Pagination
+export { PageLink }
