+++ frontend/components/SearchResults/index.ts
@@ -1,0 +1,2 @@
+import SearchResults from './SearchResults'
+export default SearchResults
