+++ frontend/components/Tabs/Tabs.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react'
+import styles from './Tabs.styles'
+
+const Tabs: React.SFC<{}> = ({ children }) => (
+  <div className={styles.tabs}>{children}</div>
+)
+
+export default Tabs
