+++ frontend/components/Tabs/index.ts
@@ -1,0 +1,5 @@
+import TabLink from './TabLink'
+import Tabs from './Tabs'
+
+export default Tabs
+export { TabLink }
