+++ frontend/components/ThemeSlider/ThemeSlider.tsx
@@ -1,8 +1,8 @@
 import { LanguageOptions, Theme } from '@vscodethemes/types'
-import { css, cx } from 'emotion'
+import { cx } from 'emotion'
 import * as React from 'react'
-import theme, { em } from '../theme'
-import { Heading, Icon, ThemePreview } from './'
+import { Heading, Icon, ThemePreview } from '../'
+import styles from './ThemeSlider.styles'
 
 interface ThemeSliderProps {
   title: string
@@ -68,7 +68,6 @@ class ThemeSlider extends React.Component<ThemeSliderProps, ThemeSliderState> {
   }
 
   next = () => {
-    const { themes } = this.props
     const { index, numOfVisibleItems } = this.state
     this.queuedIndex = index + numOfVisibleItems
     this.setState({ previousIndex: index })
@@ -111,7 +110,7 @@ class ThemeSlider extends React.Component<ThemeSliderProps, ThemeSliderState> {
       }
     }
 
-    const isAtMaxIndex = themeIndex >= maxIndex
+    // const isAtMaxIndex = themeIndex >= maxIndex
     const shouldSlide = previousIndex === index
     const isInitialSlide = shouldSlide && index === 0
     const shouldReset = !shouldSlide && index > 0
@@ -135,14 +134,14 @@ class ThemeSlider extends React.Component<ThemeSliderProps, ThemeSliderState> {
     }
 
     return (
-      <div className={classes.wrapper}>
-        <div className={classes.title}>
+      <div className={styles.wrapper}>
+        <div className={styles.title}>
           <Heading text={title} />
         </div>
-        <div className={classes.clip}>
+        <div className={styles.clip}>
           <div
             ref={this.rowRef}
-            className={classes.row}
+            className={styles.row}
             style={rowStyles}
             onTransitionEnd={this.handleSlideEnd}
           >
@@ -150,7 +149,7 @@ class ThemeSlider extends React.Component<ThemeSliderProps, ThemeSliderState> {
               <div
                 ref={i === 0 && this.itemRef}
                 key={t.themeId}
-                className={classes.item}
+                className={styles.item}
               >
                 <ThemePreview
                   theme={t}
@@ -160,94 +159,13 @@ class ThemeSlider extends React.Component<ThemeSliderProps, ThemeSliderState> {
               </div>
             ))}
           </div>
-          <button className={cx('button', classes.button)} onClick={this.next}>
-            <Icon icon="chevronRight" className={cx('icon', classes.icon)} />
+          <button className={cx('button', styles.button)} onClick={this.next}>
+            <Icon icon="chevronRight" className={cx('icon', styles.icon)} />
           </button>
         </div>
       </div>
     )
   }
-}
-
-const shadowSize = 30
-const buttonSize = [64, 128]
-const buttonOffset = theme.gutters.md
-
-const classes = {
-  wrapper: css({
-    ':hover .button': {
-      opacity: 1,
-    },
-  }),
-
-  title: css({
-    paddingLeft: em(theme.container.gutter),
-  }),
-
-  clip: css({
-    position: 'relative',
-    overflow: 'hidden',
-    // Compensate for overflow: hidden and shadow.
-    marginTop: em(-shadowSize),
-    marginBottom: em(-shadowSize),
-    padding: em(shadowSize),
-    paddingLeft: em(theme.container.gutter),
-  }),
-
-  row: css({
-    display: 'flex',
-    marginLeft: em(-theme.gutters.md / 2),
-    width: `calc(100% + ${em(theme.gutters.md / 2)})`,
-  }),
-
-  item: css({
-    // Ensure there is always a fraction of the next item visible.
-    // TODO: Use media queries to change this value.
-    width: '30%',
-    flexShrink: 0,
-    paddingLeft: em(theme.gutters.md / 2),
-    paddingRight: em(theme.gutters.md / 2),
-  }),
-
-  button: css({
-    opacity: 0,
-    position: 'absolute',
-    fontSize: 'inherit',
-    right: 0,
-    top: '50%',
-    marginTop: em(-buttonSize[1] / 2),
-    width: em(buttonSize[0] + buttonOffset),
-    height: em(buttonSize[1]),
-    paddingRight: em(buttonOffset),
-    background: `${theme.colors.backgroundDark}F0`,
-    color: theme.colors.text,
-    border: `1px solid ${theme.colors.backgroundDark}`,
-    borderRight: 'none',
-    borderTopLeftRadius: em(theme.borderRadius.md),
-    borderBottomLeftRadius: em(theme.borderRadius.md),
-    boxShadow: theme.shadows.md,
-    cursor: 'pointer',
-    transform: `translateX(${em(buttonOffset)})`,
-    transition: `opacity 0.15s ease-in, transform 0.15s ease-in`,
-
-    ':hover': {
-      transform: `translateX(0)`,
-    },
-
-    ':hover .icon': {
-      transform: `scaleY(1) translateX(${em(buttonOffset / 2)})`,
-    },
-
-    ':active .icon': {
-      transform: `scaleY(0.85) translateX(${em(buttonOffset / 2)})`,
-    },
-  }),
-
-  icon: css({
-    height: em(40),
-    transform: 'scaleY(0.85) translateX(0em)',
-    transition: 'transform 0.15s ease-in',
-  }),
 }
 
 export default ThemeSlider
