+++ frontend/components/ThemeSlider/index.ts
@@ -1,0 +1,2 @@
+import ThemeSlider from './ThemeSlider'
+export default ThemeSlider
