+++ frontend/components/index.ts
@@ -1,18 +1,12 @@
-export { default as AlgoliaLogo } from './AlgoliaLogo'
 export { default as App } from './App'
 export { default as Checkbox } from './Checkbox'
-export { default as Container } from './Container'
-export { default as Facet } from './Facet'
 export { default as Footer } from './Footer'
 export { default as Header } from './Header'
 export { default as Heading } from './Heading'
 export { default as Icon } from './Icon'
 export { default as Input } from './Input'
-export { default as Logo } from './Logo'
-export { default as PageLink } from './PageLink'
-export { default as Pagination } from './Pagination'
+export { default as Pagination, PageLink } from './Pagination'
 export { default as SearchResults } from './SearchResults'
-export { default as TabLink } from './TabLink'
-export { default as Tabs } from './Tabs'
+export { default as Tabs, TabLink } from './Tabs'
 export { default as ThemePreview } from './ThemePreview'
 export { default as ThemeSlider } from './ThemeSlider'
